refactor: split fallback job bookkeeping out of resolver_fallback
Pure code motion: the 8-function cancel/snapshot/stop tail group moves
verbatim into new resolver_fallback_jobs.py (same sibling idiom: names
resolved at call time via the resolver surface, all 8 re-exported from
resolver.py). The two same-module call sites left behind in
_start_fallback_submit_worker's closures switch to the _resolver.-prefixed
form, which the suite's patch decorators already target.

Brings resolver_fallback.py back under Codacy Lizard's 500 file-nloc
limit (507 -> ~420) ahead of the cloud gate re-enable in #386.
